+++ src/parsley/abstract.js
@@ -1,14 +1,28 @@
 define('parsley/abstract', [
-], function () {
+  'parsley/utils'
+], function (ParsleyUtils) {
   var ParsleyAbstract = function() {};
 
   ParsleyAbstract.prototype = {
     asyncSupport: false,
 
     actualizeOptions: function () {
-      this.options = this.OptionsFactory.get(this);
-
+      ParsleyUtils.attr(this.$element, this.options.namespace, this.domOptions);
+      if (this.parent)
+        this.parent.actualizeOptions();
       return this;
+    },
+
+    _resetOptions: function(initOptions) {
+      var baseOptions = this.parent ? this.parent.options : window.ParsleyConfig;
+      this.domOptions = ParsleyUtils.objectCreate(baseOptions);
+      this.options = ParsleyUtils.objectCreate(this.domOptions);
+      // Shallow copy of ownProperties of initOptions:
+      for (var i in initOptions) {
+        if (initOptions.hasOwnProperty(i))
+          this.options[i] = initOptions[i];
+      }
+      this.actualizeOptions();
     },
 
     // ParsleyValidator validate proxy function . Could be replaced by third party scripts
@@ -16,16 +30,14 @@ define('parsley/abstract', [
       return window.ParsleyValidator.validate(value, constraints, priority);
     },
 
-    // Subscribe an event and a handler for a specific field or a specific form
-    // If on a ParsleyForm instance, it will be attached to form instance and also
-    // To every field instance for this form
+    // Deprecated. Use jQuery events
     subscribe: function (name, fn) {
       $.listenTo(this, name.toLowerCase(), fn);
 
       return this;
     },
 
-    // Same as subscribe above. Unsubscribe an event for field, or form + its fields
+    // Deprecated. Use jQuery events
     unsubscribe: function (name) {
       $.unsubscribeTo(this, name.toLowerCase());
 
@@ -36,13 +48,13 @@ define('parsley/abstract', [
     reset: function () {
       // Field case: just emit a reset event for UI
       if ('ParsleyForm' !== this.__class__)
-        return $.emit('parsley:field:reset', this);
+        return this._trigger('reset');
 
       // Form case: emit a reset event for each field
       for (var i = 0; i < this.fields.length; i++)
-        $.emit('parsley:field:reset', this.fields[i]);
+        this.fields[i]._trigger('reset');
 
-      $.emit('parsley:form:reset', this);
+      this._trigger('reset');
     },
 
     // Destroy Parsley instance (+ UI)
@@ -51,7 +63,7 @@ define('parsley/abstract', [
       if ('ParsleyForm' !== this.__class__) {
         this.$element.removeData('Parsley');
         this.$element.removeData('ParsleyFieldMultiple');
-        $.emit('parsley:field:destroy', this);
+        this._trigger('destroy');
 
         return;
       }
@@ -61,7 +73,7 @@ define('parsley/abstract', [
         this.fields[i].destroy();
 
       this.$element.removeData('Parsley');
-      $.emit('parsley:form:destroy', this);
+      this._trigger('destroy');
     }
   };