+++ src/parsley/ui.js
@@ -7,16 +7,15 @@ define('parsley/ui', [
 
   ParsleyUI.prototype = {
     listen: function () {
-      $.listen('parsley:form:init', this, this.setupForm);
-      $.listen('parsley:field:init', this, this.setupField);
-
-      $.listen('parsley:field:validated', this, this.reflow);
-      $.listen('parsley:form:validated', this, this.focus);
-
-      $.listen('parsley:field:reset', this, this.reset);
-
-      $.listen('parsley:form:destroy', this, this.destroy);
-      $.listen('parsley:field:destroy', this, this.destroy);
+      var that = this;
+      $(document)
+      .on('form:init.parsley',       function(evt, field) { that.setupForm (field) } )
+      .on('field:init.parsley',      function(evt, field) { that.setupField(field) } )
+      .on('field:validated.parsley', function(evt, field) { that.reflow    (field) } )
+      .on('form:validated.parsley',  function(evt, field) { that.focus     (field) } )
+      .on('field:reset.parsley',     function(evt, field) { that.reset     (field) } )
+      .on('form:destroy.parsley',    function(evt, field) { that.destroy   (field) } )
+      .on('field:destroy.parsley',   function(evt, field) { that.destroy   (field) } );
 
       return this;
     },
@@ -64,7 +63,7 @@ define('parsley/ui', [
     },
 
     manageStatusClass: function (fieldInstance) {
-      if (true === fieldInstance.validationResult)
+      if (fieldInstance.hasConstraints() && fieldInstance.needsValidation() && true === fieldInstance.validationResult)
         this._successClass(fieldInstance);
       else if (fieldInstance.validationResult.length > 0)
         this._errorClass(fieldInstance);
@@ -79,6 +78,8 @@ define('parsley/ui', [
       // Case where we have errorMessage option that configure an unique field error message, regardless failing validators
       if ('undefined' !== typeof fieldInstance.options.errorMessage) {
         if ((diff.added.length || diff.kept.length)) {
+          this._insertErrorWrapper(fieldInstance);
+
           if (0 === fieldInstance._ui.$errorsWrapper.find('.parsley-custom-error-message').length)
             fieldInstance._ui.$errorsWrapper
               .append(
@@ -112,6 +113,7 @@ define('parsley/ui', [
     // TODO: strange API here, intuitive for manual usage with addError(pslyInstance, 'foo', 'bar')
     // but a little bit complex for above internal usage, with forced undefined parameter...
     addError: function (fieldInstance, name, message, assert, doNotUpdateClass) {
+      this._insertErrorWrapper(fieldInstance);
       fieldInstance._ui.$errorsWrapper
         .addClass('filled')
         .append(
@@ -244,12 +246,6 @@ define('parsley/ui', [
       // Store it in fieldInstance for later
       fieldInstance._ui = _ui;
 
-      // Stops excluded inputs from getting errorContainer added
-      if( !fieldInstance.$element.is(fieldInstance.options.excluded) ) {
-        /** Mess with DOM now **/
-        this._insertErrorWrapper(fieldInstance);
-      }
-
       // Bind triggers first time
       this.actualizeTriggers(fieldInstance);
     },
@@ -277,6 +273,10 @@ define('parsley/ui', [
 
     _insertErrorWrapper: function (fieldInstance) {
       var $errorsContainer;
+
+      // Nothing to do if already inserted
+      if (0 !== fieldInstance._ui.$errorsWrapper.parent().length)
+        return fieldInstance._ui.$errorsWrapper.parent();
 
       if ('string' === typeof fieldInstance.options.errorsContainer) {
         if ($(fieldInstance.options.errorsContainer).length)