+++ test/features/parsley.js
@@ -24,18 +24,17 @@ define(function () {
         expect(parsleyInstance).to.be.an('object');
         expect(parsleyInstance.__class__).to.be('ParsleyField');
       });
-      it('should return Parsley if instantiated on an unsupported element', function () {
+      it('should return ParsleyField even if instantiated on an unsupported element', function () {
         $('body').append('<div id="element"></div>');
         var parsleyInstance = new Parsley($('#element'));
         expect(parsleyInstance).to.be.an('object');
-        expect(parsleyInstance.__class__).to.be('Parsley');
+        expect(parsleyInstance.__class__).to.be('ParsleyField');
       });
-      it('should return Parsley instance if instantiated on an excluded field type, and do not have an errors container', function () {
+      it('should return ParsleyField instance even if instantiated on an excluded field type, and do not have an errors container', function () {
         $('body').append('<input type="submit" id="element" />');
         var parsleyInstance = new Parsley($('#element'));
         expect(parsleyInstance).to.be.an('object');
-        expect(parsleyInstance.__class__).to.be('Parsley');
-        expect($('#parsley-id-' + parsleyInstance.__id__).length).to.be(0);
+        expect(parsleyInstance.__class__).to.be('ParsleyField');
       });
       it('should have excluded fields by default', function () {
         $('body').append(
@@ -58,22 +57,25 @@ define(function () {
         $('body').append('<div id="element"></div>');
 
         // default ParsleyOptions.namespace
-        expect(new Parsley($('#element')).OptionsFactory.staticOptions.namespace).to.be('data-parsley-');
+        expect(new Parsley($('#element')).options.namespace).to.be('data-parsley-');
 
         // global JS config
+        $('#element').parsley().destroy()
         window.ParsleyConfig.namespace = 'data-foo-';
-        expect(new Parsley($('#element')).OptionsFactory.staticOptions.namespace).to.be('data-foo-');
+        expect(new Parsley($('#element')).options.namespace).to.be('data-foo-');
 
         // option on the go
+        $('#element').parsley().destroy()
         expect(new Parsley($('#element'), {
           namespace: "data-bar-"
-        }).OptionsFactory.staticOptions.namespace).to.be('data-bar-');
+        }).options.namespace).to.be('data-bar-');
 
         // data- DOM-API
+        $('#element').parsley().destroy()
         $('#element').attr('data-parsley-namespace', 'data-baz-');
         expect(new Parsley($('#element'), {
           namespace: "data-bar-"
-        }).OptionsFactory.staticOptions.namespace).to.be('data-baz-');
+        }).options.namespace).to.be('data-bar-');
         delete window.ParsleyConfig.namespace;
       });
       it('should handle proper options management', function () {
@@ -84,6 +86,8 @@ define(function () {
         expect(parsleyInstance.options.baz).to.be('baz');
         expect(parsleyInstance.options.bar).to.be('baz');
         expect(parsleyInstance.options.qux).to.be('bux');
+        delete ParsleyConfig.bar;
+        delete ParsleyConfig.baz;
       });
       it('should have a jquery plugin API', function () {
         $('body').append('<input type="text" id="element" data-parsley-namespace="baz-"></div>');
@@ -105,8 +109,6 @@ define(function () {
         expect($('input').parsley().length).to.be(2);
       });
       afterEach(function () {
-        window.ParsleyConfig = { i18n: window.ParsleyConfig.i18n, validators: window.ParsleyConfig.validators };
-
         $('#element').remove();
       });
     });