+++ test/features/ui.js
@@ -8,10 +8,12 @@ define(function () {
         var UI = new ParsleyUI();
         expect(UI.listen).not.to.be(undefined);
       });
-      it('should create proper errors container', function () {
+      it('should create proper errors container when needed', function () {
         $('body').append('<input type="text" id="element" data-parsley-required />');
         var parsleyField = $('#element').psly();
         expect($('#element').attr('data-parsley-id')).to.be(parsleyField.__id__);
+        expect($('ul#parsley-id-' + parsleyField.__id__).length).to.be(0);
+        parsleyField.validate();
         expect($('ul#parsley-id-' + parsleyField.__id__).length).to.be(1);
         expect($('ul#parsley-id-' + parsleyField.__id__).hasClass('parsley-errors-list')).to.be(true);
       });
@@ -22,7 +24,7 @@ define(function () {
             '<div id="container"></div>'                                                             +
             '<div id="container2"></div>'                                                            +
           '</form>');
-        $('#element').psly();
+        $('#element').psly().validate();
         expect($('#container .parsley-errors-list').length).to.be(1);
         $('#element').psly().destroy();
         $('#field1').removeAttr('data-parsley-errors-container');
@@ -32,9 +34,9 @@ define(function () {
         expect($('#container2 .parsley-errors-list').length).to.be(1);
       });
       it('should handle wrong errors-container option', function () {
-        $('body').append('<input type="text" id="element" data-parsley-errors-container="#donotexist" />');
+        $('body').append('<input type="text" id="element" data-parsley-errors-container="#donotexist" required/>');
         window.console.warn = sinon.spy();
-        var parsleyInstance = $('#element').psly();
+        $('#element').psly().validate();
         expect(window.console.warn.called).to.be(true);
       });
       it('should add proper parsley class on success or failure (type=text)', function () {
@@ -46,6 +48,20 @@ define(function () {
         $('#element').val('foo').psly().validate();
         expect($('#element').hasClass('parsley-success')).to.be(true);
         expect($('#element').hasClass('parsley-error')).to.be(false);
+      });
+      it('should not add success class on a field without constraints', function () {
+        $('body').append('<input type="text" id="element" />');
+        var parsleyField = $('#element').psly();
+        parsleyField.validate();
+        expect($('#element').hasClass('parsley-error')).to.be(false);
+        expect($('#element').hasClass('parsley-success')).to.be(false);
+      });
+      it('should not add success class on an empty optional field', function () {
+        $('body').append('<input type="number" id="element" />');
+        var parsleyField = $('#element').psly();
+        parsleyField.validate();
+        expect($('#element').hasClass('parsley-error')).to.be(false);
+        expect($('#element').hasClass('parsley-success')).to.be(false);
       });
       it('should add proper parsley class on success or failure (type=radio)', function () {
         $('body').append('<input type="radio" id="element" required />');
@@ -156,7 +172,7 @@ define(function () {
         $('#element').trigger($.Event('change'));
         expect($('ul#parsley-id-' + parsleyField.__id__ + ' li').length).to.be(1);
       });
-      it('should auto bind error trigger on selet field error (input=text)', function () {
+      it('should auto bind error trigger on select field error (input=text)', function () {
         $('body').append('<input type="email" id="element" required />');
         var parsleyField = $('#element').psly();
         expect($('ul#parsley-id-' + parsleyField.__id__ + ' li').length).to.be(0);
@@ -166,7 +182,7 @@ define(function () {
         $('#element').val('foo').trigger($.Event('keyup'));
         expect($('ul#parsley-id-' + parsleyField.__id__ + ' li').hasClass('parsley-type')).to.be(true);
       });
-      it('should auto bind error trigger on selet field error (select)', function () {
+      it('should auto bind error trigger on select field error (select)', function () {
         $('body').append('<select id="element" required>'+
           '<option value="">Choose</option>' +
           '<option value="foo">foo</option>' +